Fix syntax error and redirection
The original cron job contained an unescaped percent sign which was
apparently interpreted as a multiline-escape.
This removes the need for a percent sign alltogether by using cron's
syntax to check if we're in the first 7 days of the month.

I also changed the output redirection (which was only affecting the last
command anyways) and made wget be quiet by default.

Fixes #1